requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose
types-python-jose
passlib
python-multipart
slowapi
requests
types-requests
aiosqlite